requirements.txt
pandas>=2.0.0
numpy>=1.24.0
torch>=2.0.0
yfinance>=0.2.28
python-dotenv>=1.0.0
datasets>=2.0.0
requests>=2.31.0
openai>=1.0.0
alpaca-py>=0.10.0
transformers>=4.30.0
scikit-learn>=1.3.0
pytz>=2023.3
apify-client>=1.7.0